Finish moving update functionality out of util.py.
