+++ gslib/command_runner.py
@@ -44,10 +44,8 @@ from gslib.gcs_json_api import GcsJsonApi
 from gslib.no_op_credentials import NoOpCredentials
 from gslib.tab_complete import MakeCompleter
 from gslib.util import CheckMultiprocessingAvailableAndInit
-from gslib.util import GetGsutilVersionModifiedTime
 from gslib.util import IsCustomMetadataHeader
 from gslib.util import LAST_CHECKED_FOR_GSUTIL_UPDATE_TIMESTAMP_FILE
-from gslib.util import LookUpGsutilVersion
 from gslib.utils.constants import GSUTIL_PUB_TARBALL
 from gslib.utils.constants import RELEASE_NOTES_URL
 from gslib.utils.constants import UTF8
@@ -56,6 +54,7 @@ from gslib.utils.text_util import CompareVersions
 from gslib.utils.text_util import InsistAsciiHeader
 from gslib.utils.text_util import InsistAsciiHeaderValue
 from gslib.utils.unit_util import SECONDS_PER_DAY
+from gslib.utils.update_util import LookUpGsutilVersion
 
 
 def HandleHeaderCoding(headers):
@@ -383,7 +382,7 @@ class CommandRunner(object):
       # Set last_checked_ts from date of VERSION file, so if the user installed
       # an old copy of gsutil it will get noticed (and an update offered) the
       # first time they try to run it.
-      last_checked_ts = GetGsutilVersionModifiedTime()
+      last_checked_ts = gslib.GetGsutilVersionModifiedTime()
       with open(LAST_CHECKED_FOR_GSUTIL_UPDATE_TIMESTAMP_FILE, 'w') as f:
         f.write(str(last_checked_ts))
     else:
