+++ gslib/commands/update.py
@@ -31,7 +31,6 @@ from gslib.cs_api_map import ApiSelector
 from gslib.exception import CommandException
 from gslib.metrics import CheckAndMaybePromptForAnalyticsEnabling
 from gslib.sig_handling import RegisterSignalHandler
-from gslib.util import LookUpGsutilVersion
 from gslib.utils.boto_util import GetBotoConfigFileList
 from gslib.utils.boto_util import CERTIFICATE_VALIDATION_ENABLED
 from gslib.utils.constants import GSUTIL_PUB_TARBALL
@@ -40,6 +39,7 @@ from gslib.utils.system_util import IS_CYGWIN
 from gslib.utils.system_util import IS_WINDOWS
 from gslib.utils.text_util import CompareVersions
 from gslib.utils.update_util import DisallowUpdateIfDataInGsutilDir
+from gslib.utils.update_util import LookUpGsutilVersion
 
 
 _SYNOPSIS = """
