+++ gslib/util.py
@@ -128,37 +128,6 @@ def UnaryDictToXml(message):
   return ElementTree.tostring(element_type)
 
 
-def LookUpGsutilVersion(gsutil_api, url_str):
-  """Looks up the gsutil version of the specified gsutil tarball URL.
-
-  Version is specified in the metadata field set on that object.
-
-  Args:
-    gsutil_api: gsutil Cloud API to use when retrieving gsutil tarball.
-    url_str: tarball URL to retrieve (such as 'gs://pub/gsutil.tar.gz').
-
-  Returns:
-    Version string if URL is a cloud URL containing x-goog-meta-gsutil-version
-    metadata, else None.
-  """
-  url = StorageUrlFromString(url_str)
-  if url.IsCloudUrl():
-    obj = gsutil_api.GetObjectMetadata(url.bucket_name, url.object_name,
-                                       provider=url.scheme,
-                                       fields=['metadata'])
-    if obj.metadata and obj.metadata.additionalProperties:
-      for prop in obj.metadata.additionalProperties:
-        if prop.key == 'gsutil_version':
-          return prop.value
-
-
-def GetGsutilVersionModifiedTime():
-  """Returns unix timestamp of when the VERSION file was last modified."""
-  if not gslib.VERSION_FILE:
-    return 0
-  return int(os.path.getmtime(gslib.VERSION_FILE))
-
-
 def CreateCustomMetadata(entries=None, custom_metadata=None):
   """Creates a custom metadata (apitools Object.MetadataValue) object.
