+++ gslib/utils/update_util.py
@@ -21,6 +21,7 @@ import textwrap
 
 import gslib
 from gslib.exception import CommandException
+from gslib.storage_url import StorageUrlFromString
 
 
 # This function used to belong inside of update.py. However, it needed to be
@@ -79,5 +80,25 @@ def DisallowUpdateIfDataInGsutilDir(directory=gslib.GSUTIL_DIR):
           os.path.join(gslib.GSUTIL_DIR, filename))))
 
 
+def LookUpGsutilVersion(gsutil_api, url_str):
+  """Looks up the gsutil version of the specified gsutil tarball URL.
 
+  Version is specified in the metadata field set on that object.
 
+  Args:
+    gsutil_api: gsutil Cloud API to use when retrieving gsutil tarball.
+    url_str: tarball URL to retrieve (such as 'gs://pub/gsutil.tar.gz').
+
+  Returns:
+    Version string if URL is a cloud URL containing x-goog-meta-gsutil-version
+    metadata, else None.
+  """
+  url = StorageUrlFromString(url_str)
+  if url.IsCloudUrl():
+    obj = gsutil_api.GetObjectMetadata(url.bucket_name, url.object_name,
+                                       provider=url.scheme,
+                                       fields=['metadata'])
+    if obj.metadata and obj.metadata.additionalProperties:
+      for prop in obj.metadata.additionalProperties:
+        if prop.key == 'gsutil_version':
+          return prop.value
